Rename pages to components
