+++ src/components/connection-list-item.jsx
