+++ src/containers/connection.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router';
 import * as DatabaseActions from '../actions/databases.js';
 import * as QueryActions from '../actions/query.js';
-import DatabaseList from '../pages/database-list.jsx';
-import Database from '../pages/database.jsx';
+import DatabaseList from '../components/database-list.jsx';
+import Database from '../components/database.jsx';
 
 
 const STYLES = {
